+++ apps/meteor/app/notifications/client/lib/Presence.ts
@@ -1,17 +1,15 @@
-import type { StreamerEvents } from '@rocket.chat/ui-contexts';
+import type { UserStatus } from '@rocket.chat/core-typings';
 import { Meteor } from 'meteor/meteor';
 
-import { Presence, STATUS_MAP } from '../../../../client/lib/presence';
+import { Presence } from '../../../../client/lib/presence';
 
 // TODO implement API on Streamer to be able to listen to all streamed data
 // this is a hacky way to listen to all streamed data from user-presence Streamer
 
 new Meteor.Streamer('user-presence');
 
-(Meteor as any).StreamerCentral.on('stream-user-presence', (uid: string, ...args: StreamerEvents['user-presence'][number]['args']) => {
-	if (!Array.isArray(args)) {
-		throw new Error('Presence event must be an array');
-	}
-	const [[username, status, statusText]] = args;
-	Presence.notify({ _id: uid, username, status: STATUS_MAP[status ?? 0], statusText });
+type args = [username: string, statusChanged?: UserStatus, statusText?: string];
+
+Meteor.StreamerCentral.on('stream-user-presence', (uid: string, [username, statusChanged, statusText]: args) => {
+	Presence.notify({ _id: uid, username, status: statusChanged, statusText });
 });