+++ apps/meteor/client/views/cloud/CloudAnnouncementHandler.tsx
@@ -1,4 +1,5 @@
-import type { IBanner, UiKit } from '@rocket.chat/core-typings';
+import type { IBanner } from '@rocket.chat/core-typings';
+import type * as UiKit from '@rocket.chat/ui-kit';
 import { useEffect, useRef } from 'react';
 import { useTranslation } from 'react-i18next';