+++ apps/meteor/client/views/omnichannel/triggers/EditTrigger.tsx
@@ -4,7 +4,6 @@ import {
 	FieldGroup,
 	Button,
 	ButtonGroup,
-	Box,
 	Field,
 	FieldLabel,
 	FieldRow,
@@ -13,6 +12,7 @@ import {
 	ToggleSwitch,
 	Select,
 	TextAreaInput,
+	NumberInput,
 } from '@rocket.chat/fuselage';
 import { useUniqueId } from '@rocket.chat/fuselage-hooks';
 import { useToastMessageDispatch, useRouter, useEndpoint, useTranslation } from '@rocket.chat/ui-contexts';
@@ -155,28 +155,24 @@ const EditTrigger = ({ triggerData }: { triggerData?: Serialized<ILivechatTrigge
 				<form id={formId} onSubmit={handleSubmit(handleSave)}>
 					<FieldGroup>
 						<Field>
-							<Box display='flex' flexDirection='row'>
+							<FieldRow>
 								<FieldLabel htmlFor={enabledField}>{t('Enabled')}</FieldLabel>
-								<FieldRow>
-									<Controller
-										name='enabled'
-										control={control}
-										render={({ field: { value, ...field } }) => <ToggleSwitch id={enabledField} {...field} checked={value} />}
-									/>
-								</FieldRow>
-							</Box>
+								<Controller
+									name='enabled'
+									control={control}
+									render={({ field: { value, ...field } }) => <ToggleSwitch id={enabledField} {...field} checked={value} />}
+								/>
+							</FieldRow>
 						</Field>
 						<Field>
-							<Box display='flex' flexDirection='row'>
+							<FieldRow>
 								<FieldLabel htmlFor={runOnceField}>{t('Run_only_once_for_each_visitor')}</FieldLabel>
-								<FieldRow>
-									<Controller
-										name='runOnce'
-										control={control}
-										render={({ field: { value, ...field } }) => <ToggleSwitch id={runOnceField} {...field} checked={value} />}
-									/>
-								</FieldRow>
-							</Box>
+								<Controller
+									name='runOnce'
+									control={control}
+									render={({ field: { value, ...field } }) => <ToggleSwitch id={runOnceField} {...field} checked={value} />}
+								/>
+							</FieldRow>
 						</Field>
 						<Field>
 							<FieldLabel htmlFor={nameField} required>
@@ -232,7 +228,13 @@ const EditTrigger = ({ triggerData }: { triggerData?: Serialized<ILivechatTrigge
 											<Controller
 												name={`conditions.${index}.value`}
 												control={control}
-												render={({ field }) => <TextInput {...field} placeholder={conditionValuePlaceholder} />}
+												render={({ field }) => {
+													if (conditions[index].name === 'time-on-site') {
+														return <NumberInput {...field} placeholder={conditionValuePlaceholder} />;
+													}
+
+													return <TextInput {...field} placeholder={conditionValuePlaceholder} />;
+												}}
 											/>
 										</FieldRow>
 									)}