+++ apps/meteor/client/views/room/modals/ReactionListModal/ReactionUserTag.tsx
@@ -1,18 +1,30 @@
 import type { IUser } from '@rocket.chat/core-typings';
 import { Box, Tag } from '@rocket.chat/fuselage';
-import type { ReactElement } from 'react';
+import { useEffectEvent } from '@rocket.chat/fuselage-hooks';
+import type { ReactElement, UIEvent } from 'react';
 import React from 'react';
 
 type ReactionUserTagProps = {
 	username: IUser['username'];
-	onClick: (e: React.MouseEvent<HTMLElement>) => void;
 	displayName: string;
+	onOpenUserCard?: (e: UIEvent, username: string) => void;
 };
 
-const ReactionUserTag = ({ username, onClick, displayName }: ReactionUserTagProps): ReactElement => (
-	<Box mie={4} mbe={4} data-username={username} onClick={onClick} key={displayName}>
-		<Tag variant='primary'>{displayName}</Tag>
-	</Box>
-);
+const ReactionUserTag = ({ username, displayName, onOpenUserCard }: ReactionUserTagProps): ReactElement => {
+	const handleOpenCard = useEffectEvent((e: UIEvent) => {
+		if (!username) {
+			return;
+		}
 
+		onOpenUserCard?.(e, username);
+	});
+
+	return (
+		<Box mie={4} mbe={4}>
+			<Tag onClick={handleOpenCard} variant='primary'>
+				{displayName}
+			</Tag>
+		</Box>
+	);
+};
 export default ReactionUserTag;