+++ apps/meteor/server/models/raw/LivechatDepartmentAgents.ts
@@ -78,6 +78,10 @@ export class LivechatDepartmentAgentsRaw extends BaseRaw<ILivechatDepartmentAgen
 		return this.find(query, options);
 	}
 
+	findByAgentIds(agentIds: string[], options?: FindOptions<ILivechatDepartmentAgents>): FindCursor<ILivechatDepartmentAgents> {
+		return this.find({ agentId: { $in: agentIds } }, options);
+	}
+
 	findByAgentId(agentId: string, options?: FindOptions<ILivechatDepartmentAgents>): FindCursor<ILivechatDepartmentAgents> {
 		return this.find({ agentId }, options);
 	}
@@ -238,7 +242,7 @@ export class LivechatDepartmentAgentsRaw extends BaseRaw<ILivechatDepartmentAgen
 			},
 		};
 
-		const agent = await this.col.findOneAndUpdate(query, update, { sort, returnDocument: 'after' });
+		const agent = await this.findOneAndUpdate(query, update, { sort, returnDocument: 'after' });
 		if (agent?.value) {
 			return {
 				agentId: agent.value.agentId,
@@ -336,7 +340,7 @@ export class LivechatDepartmentAgentsRaw extends BaseRaw<ILivechatDepartmentAgen
 			},
 		};
 
-		const bot = await this.col.findOneAndUpdate(query, update, { sort, returnDocument: 'after' });
+		const bot = await this.findOneAndUpdate(query, update, { sort, returnDocument: 'after' });
 		if (bot?.value) {
 			return {
 				agentId: bot.value.agentId,