+++ apps/meteor/server/models/raw/Users.js
@@ -1699,7 +1699,7 @@ export class UsersRaw extends BaseRaw {
 			},
 		};
 
-		const user = await this.col.findOneAndUpdate(query, update, { sort, returnDocument: 'after' });
+		const user = await this.findOneAndUpdate(query, update, { sort, returnDocument: 'after' });
 		if (user && user.value) {
 			return {
 				agentId: user.value._id,
@@ -1729,7 +1729,7 @@ export class UsersRaw extends BaseRaw {
 			},
 		};
 
-		const user = await this.col.findOneAndUpdate(query, update, { sort, returnDocument: 'after' });
+		const user = await this.findOneAndUpdate(query, update, { sort, returnDocument: 'after' });
 		if (user?.value) {
 			return {
 				agentId: user.value._id,
@@ -1984,45 +1984,63 @@ export class UsersRaw extends BaseRaw {
 		);
 	}
 
-	removeExpiredEmailCodesOfUserId(userId) {
+	removeExpiredEmailCodeOfUserId(userId) {
+		return this.updateOne(
+			{ '_id': userId, 'services.emailCode.expire': { $lt: new Date() } },
+			{
+				$unset: { 'services.emailCode': 1 },
+			},
+		);
+	}
+
+	removeEmailCodeOfUserId(userId) {
 		return this.updateOne(
 			{ _id: userId },
 			{
-				$pull: {
-					'services.emailCode': {
-						expire: { $lt: new Date() },
-					},
+				$unset: { 'services.emailCode': 1 },
+			},
+		);
+	}
+
+	incrementInvalidEmailCodeAttempt(userId) {
+		return this.findOneAndUpdate(
+			{ _id: userId },
+			{
+				$inc: { 'services.emailCode.attempts': 1 },
+			},
+			{
+				returnDocument: 'after',
+				projection: {
+					'services.emailCode.attempts': 1,
 				},
 			},
 		);
 	}
 
-	removeEmailCodeByUserIdAndCode(userId, code) {
-		return this.updateOne(
-			{ _id: userId },
+	async maxInvalidEmailCodeAttemptsReached(userId, maxAttempts) {
+		const result = await this.findOne(
 			{
-				$pull: {
-					'services.emailCode': {
-						code,
-					},
+				'_id': userId,
+				'services.emailCode.attempts': { $gte: maxAttempts },
+			},
+			{
+				projection: {
+					_id: 1,
 				},
 			},
 		);
+		return !!result?._id;
 	}
 
 	addEmailCodeByUserId(userId, code, expire) {
 		return this.updateOne(
 			{ _id: userId },
 			{
-				$push: {
+				$set: {
 					'services.emailCode': {
-						$each: [
-							{
-								code,
-								expire,
-							},
-						],
-						$slice: -5,
+						code,
+						expire,
+						attempts: 0,
 					},
 				},
 			},