+++ apps/meteor/tests/e2e/utils/omnichannel/rooms.ts
@@ -4,9 +4,11 @@ import { BaseTest } from '../test';
 
 type UpdateRoomParams = { roomId: string; visitorId: string; tags: string[] };
 
+type CloseRoomParams = { roomId: string; visitorToken: string };
+
 type CreateRoomParams = { tags?: string[]; visitorToken: string; agentId?: string };
 
-type CreateVisitorParams = { token: string; department?: string; name: string; email: string };
+type CreateVisitorParams = { token: string; departmentId?: string; name?: string; email?: string };
 
 type CreateConversationParams = { visitorName?: string; visitorToken?: string; agentId?: string; departmentId?: string };
 
@@ -19,11 +21,20 @@ export const updateRoom = async (api: BaseTest['api'], { roomId, visitorId, tags
 		throw Error('Unable to update room info, missing visitor id');
 	}
 
-	return api.post('/livechat/room.saveInfo', {
+	const response = await api.post('/livechat/room.saveInfo', {
 		guestData: { _id: visitorId },
 		roomData: { _id: roomId, tags },
 	});
+
+	if (response.status() !== 200) {
+		throw Error(`Unable to update room info [http status: ${response.status()}]`);
+	}
+
+	return response;
 };
+
+export const closeRoom = async (api: BaseTest['api'], { roomId, visitorToken }: CloseRoomParams) =>
+	api.post('/livechat/room.close', { rid: roomId, token: visitorToken });
 
 export const createRoom = async (api: BaseTest['api'], { visitorToken, agentId }: CreateRoomParams) => {
 	const response = await api.get('/livechat/room', {
@@ -32,69 +43,86 @@ export const createRoom = async (api: BaseTest['api'], { visitorToken, agentId }
 	});
 
 	if (response.status() !== 200) {
-		throw Error(`Unable to create room [http status: ${response.status()}]`);
+		throw Error(`Unable to create room [http status: ${response.status()}]`, { cause: await response.json() });
 	}
 
-	const data = await response.json();
+	const { room } = await response.json();
 
 	return {
 		response,
-		data: data.room,
-		delete: async () => {
-			await api.post('/livechat/room.close', { rid: data.room._id, token: visitorToken });
-			await api.post('/method.call/livechat:removeRoom', {
+		data: room,
+		async delete() {
+			await closeRoom(api, { roomId: room._id, visitorToken });
+			return api.post('/method.call/livechat:removeRoom', {
 				message: JSON.stringify({
 					msg: 'method',
 					id: '16',
 					method: 'livechat:removeRoom',
-					params: [data.room._id],
+					params: [room._id],
 				}),
 			});
 		},
 	};
 };
 
-export const createVisitor = async (api: BaseTest['api'], params: CreateVisitorParams) =>
-	api.post('/livechat/visitor', { visitor: params });
+export const createVisitor = async (api: BaseTest['api'], { name, token, departmentId }: CreateVisitorParams) => {
+	const response = await api.post('/livechat/visitor', {
+		visitor: {
+			name: name || faker.person.fullName(),
+			email: faker.internet.email(),
+			token,
+			...(departmentId && { department: departmentId }),
+		},
+	});
+
+	if (response.status() !== 200) {
+		throw Error(`Unable to create visitor [http status: ${response.status()}]`);
+	}
+
+	const { visitor } = await response.json();
+
+	return {
+		response,
+		data: visitor,
+		delete: async () => api.delete(`/livechat/visitor/${token}`),
+	};
+};
 
 export const sendMessageToRoom = async (
 	api: BaseTest['api'],
 	{ visitorToken, roomId, message }: { visitorToken: string; roomId: string; message?: string },
-) =>
-	api.post(`/livechat/message`, {
+) => {
+	const response = await api.post(`/livechat/message`, {
 		token: visitorToken,
 		rid: roomId,
 		msg: message || faker.lorem.sentence(),
 	});
 
+	if (response.status() !== 200) {
+		throw Error(`Unable to send message to room [http status: ${response.status()}]`);
+	}
+
+	return response;
+};
+
 export const createConversation = async (
 	api: BaseTest['api'],
-	{ visitorName, visitorToken, agentId, departmentId }: CreateConversationParams,
+	{ visitorName, visitorToken, agentId, departmentId }: CreateConversationParams = {},
 ) => {
 	const token = visitorToken || faker.string.uuid();
-	const visitorRes = await createVisitor(api, {
-		name: visitorName || faker.person.firstName(),
-		email: faker.internet.email(),
-		token,
-		...(departmentId && { department: departmentId }),
-	});
 
-	if (visitorRes.status() !== 200) {
-		throw Error(`Unable to create visitor [http status: ${visitorRes.status()}]`);
-	}
-
-	const { data: room } = await createRoom(api, { visitorToken: token, agentId });
-
-	const messageRes = await sendMessageToRoom(api, { visitorToken: token, roomId: room._id });
-
-	if (messageRes.status() !== 200) {
-		throw Error(`Unable to send message to room [http status: ${messageRes.status()}]`);
-	}
-
-	const { visitor } = await visitorRes.json();
+	const { data: visitor, delete: deleteVisitor } = await createVisitor(api, { token, name: visitorName, departmentId });
+	const { data: room, delete: deleteRoom } = await createRoom(api, { visitorToken: token, agentId });
+	await sendMessageToRoom(api, { visitorToken: token, roomId: room._id });
 
 	return {
-		room,
-		visitor,
+		data: {
+			room,
+			visitor,
+		},
+		delete: async () => {
+			await deleteRoom();
+			await deleteVisitor();
+		},
 	};
 };