+++ apps/meteor/tests/end-to-end/api/03-groups.js
@@ -1264,7 +1264,7 @@ describe('[Groups]', function () {
 	});
 
 	describe('/groups.files', async () => {
-		await testFileUploads('groups.files', group);
+		await testFileUploads('groups.files', 'p');
 	});
 
 	describe('/groups.listAll', () => {