+++ apps/meteor/ee/server/services/CHANGELOG.md
@@ -1,5 +1,158 @@
 # rocketchat-services
 
+## 1.1.22
+
+### Patch Changes
+
+- ([#31138](https://github.com/RocketChat/Rocket.Chat/pull/31138)) feat(uikit): Move `@rocket.chat/ui-kit` package to the main monorepo
+
+- ([#31349](https://github.com/RocketChat/Rocket.Chat/pull/31349) by [@Subhojit-Dey1234](https://github.com/Subhojit-Dey1234)) feat: Implemented InlineCode handling in Bold, Italic and Strike
+
+- <details><summary>Updated dependencies [b223cbde14, dbb08ef948, fae558bd5d, 748e57984d, 7c6198f49f, 9a6e9b4e28, fdd9852079, 2260c04ec6, c8ab6583dc, e7d3cdeef0, b4b2cd20a8]:</summary>
+
+  - @rocket.chat/ui-kit@0.33.0
+  - @rocket.chat/core-services@0.3.4
+  - @rocket.chat/model-typings@0.3.0
+  - @rocket.chat/core-typings@6.6.0
+  - @rocket.chat/rest-typings@6.6.0
+  - @rocket.chat/models@0.0.28
+  </details>
+
+## 1.1.22-rc.7
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/core-typings@6.6.0-rc.7
+  - @rocket.chat/rest-typings@6.6.0-rc.7
+  - @rocket.chat/core-services@0.3.4-rc.7
+  - @rocket.chat/model-typings@0.3.0-rc.7
+  - @rocket.chat/models@0.0.28-rc.7
+  </details>
+
+## 1.1.22-rc.6
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/core-typings@6.6.0-rc.6
+  - @rocket.chat/rest-typings@6.6.0-rc.6
+  - @rocket.chat/core-services@0.3.4-rc.6
+  - @rocket.chat/model-typings@0.3.0-rc.6
+  - @rocket.chat/models@0.0.28-rc.6
+  </details>
+
+## 1.1.22-rc.5
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/core-typings@6.6.0-rc.5
+  - @rocket.chat/rest-typings@6.6.0-rc.5
+  - @rocket.chat/core-services@0.3.4-rc.5
+  - @rocket.chat/model-typings@0.3.0-rc.5
+  - @rocket.chat/models@0.0.28-rc.5
+  </details>
+
+## 1.1.22-rc.4
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.6.0-rc.4
+- @rocket.chat/rest-typings@6.6.0-rc.4
+- @rocket.chat/core-services@0.3.4-rc.4
+- @rocket.chat/model-typings@0.3.0-rc.4
+- @rocket.chat/models@0.0.28-rc.4
+
+## 1.1.22-rc.3
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.6.0-rc.3
+- @rocket.chat/rest-typings@6.6.0-rc.3
+- @rocket.chat/core-services@0.3.4-rc.3
+- @rocket.chat/model-typings@0.3.0-rc.3
+- @rocket.chat/models@0.0.28-rc.3
+
+## 1.1.22-rc.2
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.6.0-rc.2
+- @rocket.chat/rest-typings@6.6.0-rc.2
+- @rocket.chat/core-services@0.3.4-rc.2
+- @rocket.chat/model-typings@0.3.0-rc.2
+- @rocket.chat/models@0.0.28-rc.2
+
+## 1.1.22-rc.1
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.6.0-rc.1
+- @rocket.chat/rest-typings@6.6.0-rc.1
+- @rocket.chat/core-services@0.3.4-rc.1
+- @rocket.chat/model-typings@0.3.0-rc.1
+- @rocket.chat/models@0.0.28-rc.1
+
+## 1.1.22-rc.0
+
+### Patch Changes
+
+- b223cbde14: feat(uikit): Move `@rocket.chat/ui-kit` package to the main monorepo
+- dbb08ef948: feat: Implemented InlineCode handling in Bold, Italic and Strike
+- Updated dependencies [b223cbde14]
+- Updated dependencies [dbb08ef948]
+- Updated dependencies [fae558bd5d]
+- Updated dependencies [748e57984d]
+- Updated dependencies [7c6198f49f]
+- Updated dependencies [9a6e9b4e28]
+- Updated dependencies [fdd9852079]
+- Updated dependencies [2260c04ec6]
+- Updated dependencies [c8ab6583dc]
+- Updated dependencies [e7d3cdeef0]
+- Updated dependencies [b4b2cd20a8]
+  - @rocket.chat/ui-kit@0.33.0-rc.0
+  - @rocket.chat/core-services@0.3.4-rc.0
+  - @rocket.chat/model-typings@0.3.0-rc.0
+  - @rocket.chat/core-typings@6.6.0-rc.0
+  - @rocket.chat/rest-typings@6.6.0-rc.0
+  - @rocket.chat/models@0.0.28-rc.0
+
+## 1.1.21
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.5.3
+- @rocket.chat/rest-typings@6.5.3
+- @rocket.chat/core-services@0.3.3
+- @rocket.chat/model-typings@0.2.3
+- @rocket.chat/models@0.0.27
+
+## 1.1.20
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.5.2
+- @rocket.chat/rest-typings@6.5.2
+- @rocket.chat/core-services@0.3.2
+- @rocket.chat/model-typings@0.2.2
+- @rocket.chat/models@0.0.26
+
+## 1.1.19
+
+### Patch Changes
+
+- Updated dependencies [c2b224fd82]
+- Updated dependencies [c2b224fd82]
+  - @rocket.chat/rest-typings@6.5.1
+  - @rocket.chat/core-typings@6.5.1
+  - @rocket.chat/core-services@0.3.1
+  - @rocket.chat/model-typings@0.2.1
+  - @rocket.chat/models@0.0.25
+
 ## 1.1.18
 
 ### Patch Changes