+++ packages/ui-client/CHANGELOG.md
@@ -1,5 +1,110 @@
 # @rocket.chat/ui-client
 
+## 4.0.0
+
+### Minor Changes
+
+- ([#31516](https://github.com/RocketChat/Rocket.Chat/pull/31516)) Room header keyboard navigability
+
+  ![Kapture 2024-01-22 at 11 33 14](https://github.com/RocketChat/Rocket.Chat/assets/27704687/f116c1e6-4ec7-4175-a01b-fa98eade2416)
+
+- ([#30680](https://github.com/RocketChat/Rocket.Chat/pull/30680)) feat: Skip to main content shortcut and useDocumentTitle
+
+### Patch Changes
+
+- ([#31138](https://github.com/RocketChat/Rocket.Chat/pull/31138)) feat(uikit): Move `@rocket.chat/ui-kit` package to the main monorepo
+
+- <details><summary>Updated dependencies [b223cbde14, b2b0035162]:</summary>
+
+  - @rocket.chat/ui-contexts@4.0.0
+  </details>
+
+## 4.0.0-rc.7
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/ui-contexts@4.0.0-rc.7
+  </details>
+
+## 4.0.0-rc.6
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/ui-contexts@4.0.0-rc.6
+  </details>
+
+## 4.0.0-rc.5
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/ui-contexts@4.0.0-rc.5
+  </details>
+
+## 4.0.0-rc.4
+
+### Patch Changes
+
+- @rocket.chat/ui-contexts@4.0.0-rc.4
+
+## 4.0.0-rc.3
+
+### Patch Changes
+
+- @rocket.chat/ui-contexts@4.0.0-rc.3
+
+## 4.0.0-rc.2
+
+### Patch Changes
+
+- @rocket.chat/ui-contexts@4.0.0-rc.2
+
+## 4.0.0-rc.1
+
+### Patch Changes
+
+- @rocket.chat/ui-contexts@4.0.0-rc.1
+
+## 4.0.0-rc.0
+
+### Minor Changes
+
+- 9cb97965ba: Room header keyboard navigability
+
+  ![Kapture 2024-01-22 at 11 33 14](https://github.com/RocketChat/Rocket.Chat/assets/27704687/f116c1e6-4ec7-4175-a01b-fa98eade2416)
+
+- dd5fd6d2c8: feat: Skip to main content shortcut and useDocumentTitle
+
+### Patch Changes
+
+- b223cbde14: feat(uikit): Move `@rocket.chat/ui-kit` package to the main monorepo
+- Updated dependencies [b223cbde14]
+- Updated dependencies [b2b0035162]
+  - @rocket.chat/ui-contexts@4.0.0-rc.0
+
+## 3.0.3
+
+### Patch Changes
+
+- @rocket.chat/ui-contexts@3.0.3
+
+## 3.0.2
+
+### Patch Changes
+
+- @rocket.chat/ui-contexts@3.0.2
+
+## 3.0.1
+
+### Patch Changes
+
+- @rocket.chat/ui-contexts@3.0.1
+
 ## 3.0.0
 
 ### Patch Changes